+++ CHANGELOG.md
@@ -37,6 +37,7 @@ Compatibility:
 * Add `Range#reverse_each` (#3681, @andrykonchin).
 * Emit a warning when `it` call without arguments is used in a block without parameters (#3681, @andrykonchin).
 * Add `rb_syserr_fail_str()` (#3732, @andrykonchin).
+* Add `Dir.for_fd` (#3681, @andrykonchin).
 
 Performance: